we-move-save-get-move: Move weight save / get_weights_by_name onto WeightExporter (cut+paste)
Refactor chain ID: we-move-save-get-move
