+++ python/sglang/srt/managers/scheduler_components/weight_updater.py
@@ -291,26 +291,17 @@ class SchedulerWeightUpdaterManager:
     def save_remote_model(self, params):
         url = params["url"]
 
-        from sglang.srt.model_executor.model_runner import ModelRunner
-
-        ModelRunner.save_remote_model(self.tp_worker.model_runner.weight_exporter, url)
+        self.tp_worker.model_runner.weight_exporter.save_remote_model(url)
 
         if self.draft_worker is not None:
             draft_url = params.get("draft_url", None)
             assert (
                 draft_url is not None
             ), "draft_url must be provided when draft model is enabled"
-            from sglang.srt.model_executor.model_runner import ModelRunner
-
-            ModelRunner.save_remote_model(
-                self.draft_worker.model_runner.weight_exporter, draft_url
-            )
+            self.draft_worker.model_runner.weight_exporter.save_remote_model(draft_url)
 
     def save_sharded_model(self, params):
-        from sglang.srt.model_executor.model_runner import ModelRunner
-
-        ModelRunner.save_sharded_model(
-            self.tp_worker.model_runner.weight_exporter,
+        self.tp_worker.model_runner.weight_exporter.save_sharded_model(
             path=params["path"],
             pattern=params["pattern"],
             max_size=params["max_size"],
