+++ python/sglang/srt/model_executor/model_runner_components/weight_exporter.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import logging
 from dataclasses import dataclass, field
-from typing import Any
+from typing import Any, Optional
 
 import torch
 import torch.distributed as dist
@@ -119,3 +119,41 @@ class WeightExporter:
         torch.distributed.distributed_c10d.destroy_process_group(send_group)
         current_platform.empty_cache()
         return success, message
+
+    def save_remote_model(self: WeightExporter, url: str):
+        from sglang.srt.model_loader.loader import RemoteModelLoader
+
+        logger.info(f"Saving model to {url}")
+        RemoteModelLoader.save_model(
+            self._mr.model, self._mr.model_config.model_path, url
+        )
+
+    def save_sharded_model(
+        self: WeightExporter,
+        path: str,
+        pattern: Optional[str] = None,
+        max_size: Optional[int] = None,
+    ):
+        from sglang.srt.model_loader.loader import ShardedStateLoader
+
+        logger.info(
+            f"Save sharded model to {path} with pattern {pattern} and max_size {max_size}"
+        )
+        ShardedStateLoader.save_model(self._mr.model, path, pattern, max_size)
+
+    def get_weights_by_name(
+        self: WeightExporter, name: str, truncate_size: int = 100
+    ) -> Optional[torch.Tensor]:
+        """Get the weights of the parameter by its name. Similar to `get_parameter` in Hugging Face.
+
+        Only used for unit test with an unoptimized performance.
+        For optimized performance, please use torch.save and torch.load.
+        """
+        # TODO: (chenyang) Add support for Qwen models.
+        try:
+            return self._mr.model.get_weights_by_name(
+                name, truncate_size, tp_size=self._mr.tp_size
+            )
+        except Exception as e:
+            logger.error(f"Error when getting parameter {name}: {e}")
+            return None
